// web/src/schemas/public/Jobs.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.jobs */
export type JobsId = number & { __brand: 'JobsId' };

/** Represents the table public.jobs */
export default interface Jobs {
  id: JobsId;

  owner: string;

  object_id: string;

  status: string;

  created_at: Date | null;

  last_edited_time: Date | null;

  title: string | null;
}

/** Represents the initializer for the table public.jobs */
export interface JobsInitializer {
  /** Default value: nextval('jobs_id_seq'::regclass) */
  id?: JobsId;

  owner: string;

  object_id: string;

  /** Default value: 'started'::character varying */
  status?: string;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date | null;

  /** Default value: CURRENT_TIMESTAMP */
  last_edited_time?: Date | null;

  title?: string | null;
}

/** Represents the mutator for the table public.jobs */
export interface JobsMutator {
  id?: JobsId;

  owner?: string;

  object_id?: string;

  status?: string;

  created_at?: Date | null;

  last_edited_time?: Date | null;

  title?: string | null;
}